Lab06 final
checked htmls right before turning in and very glad i did. sorry for all the pushes to github - but there have been some small details with exporting code to source file and knitting.
there were a lot of uses of d2 as a dataset which i tried to replace with uniquely identified datasets.